Minor fixes, freeze code here for now
A copy of this code was made to test RayStation 6. If possible, don't
change anything for a while, in hopes that we'll be able to use the
updated scripts for RS 6 without having to track all the changes done to
the 4.7.2 scripts in the interim. I know, beau rêve.